# src/routes/api-doc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Code Challenge API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/countries": {
            "get": {
                "operationId": "getCountries",
                "summary": "Get a list of countries",
                "parameters": [
                    {
                        "name": "region",
                        "in": "query",
                        "description": "Filter countries by region",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "name": "Turkey",
                                        "region": "EMEA"
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/salesrep": {
            "get": {
                "operationId": "analyzeSalesRep",
                "summary": "Analyze regions and determine sales representative requirements",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Sales representative analysis completed successfully for the specified regions"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/optimal": {
            "get": {
                "operationId": "distributeOptimally",
                "summary": "Distribute countries to sales representatives optimally",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Optimal distribution of countries to sales representatives completed successfully",
                                    "roster": [
                                        {
                                            "region": "EMEA",
                                            "countryList": [
                                                "Turkey",
                                                "France",
                                                "Italy"
                                            ],
                                            "countryCount": 3
                                        },
                                        {
                                            "region": "APAC",
                                            "countryList": [
                                                "Japan",
                                                "China",
                                                "Australia"
                                            ],
                                            "countryCount": 3
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}